move images to assets folder and slight changes
